Add new attacking system where melee is combined with AOE
New attacking system involves a resource bar that charges over time and is used to power an area-of-effect (AOE) attack. Melee attacking does no damage and is solely used for keeping enemies at a distance. Bonus resource recovery for the AOE attack is rewarded if hitting enemies with a melee attack that started on a beat scroller hit.

Known issues:
-Enemies spawned after level load appear to not work, making surprise arenas impossible; make sure all enemies in the level are active on scene load until this is solved.
-An issue is causing arena zones to not open the gates after all arena enemies have been defeated.