Pack UI into prefab and add option to exit to title screen in-game
The user interface (UI) is now packed as a prefab for portability, allowing it to be modified without editing the level. Additionally, the player can now exit to the title screen from in-game by pressing the escape key. Minor changes were made to melee attack cooldown so players can follow the rhythm with a smaller margin of error and lure radius increased so that enemies can be attracted from across large pitfalls.