Revert prefabs in the level for portability
Reverted most values in several prefab instances in the level. These prefab instances should now read from their base prefab instead of scene overrides. This should reduce the risk of scene conflicts, so long as edits to prefab variables are done in the prefabs instead of the scene whenever possible.