Add instrument tutorial to level
Instrument text tutorial has been set up in the main level. Reactivated the enhanced melee attack and cleaned up some code in the player's character controller script.

Fixed pause menu being visible by default.
Applied a potential fix for disappearing enemies. Reoccurrence of the bug has yet to be seen.

There is a known issue where the player does not properly lock in place when the tutorial step is set to do so.