Add camera modifier zone near puzzle door
Added a trigger zone that modifies camera position and rotation and placed one in the level in front of the puzzle door. Moved puzzle pieces so that the camera can fit all important objects in frame when the camera is at the modified angle and so that the player will pick up the first statue before they touch the statue bases at the door (this can be changed back later if the puzzle flows better with the old arrangement).

There is a known issue where the angle of view shown in the camera modifier zone does not match the angle applied during gameplay.