Add context-sensitive movement tutorial
Extended the versatility of the TriggerTutorial script so that it can also activate after a delay if the player remains in the trigger area. The newly-added movement tutorial uses this new code to activate the tutorial only if the player doesn't leave the beach after a set period of time.