Add finishing touches to door puzzle
Added pulsing movement to door puzzle instruments when they make a sound for extra visual feedback. Made instruments for the door puzzle go back to their initial locations when the player respawns to prevent them from going out of bounds. This also encourages the player to return instruments to their pedestals so they do not lose them (and discourages players from "cheating" the puzzle). Additionally, added an extra null-reference check in the player's melee attack loop. This is intended to prevent the script from crashing when attacking enemies that are being destroyed by other processes (which may be in an unsynced update loop).