Fix bug applying invulnerability time on healing
Added a conditional check so that the basic apply damage call (the one without knockback) no longer makes the player invulnerable to healing for a short timeframe.
Note that calls for damage with knockback should not be used for healing, as there is no code supporting this circumstance and the ability to do so would require code for separate invulnerability timers for knockback and damage, which would complicate further additions to the damage/knockback code.