Make player's tunic flash when health changes
When gaining or losing health, the player character's tunic flashes either red or green to indicate damage and healing respectively. The change in colour is not very noticeable, so it might be beneficial in the future to add an emissive effect to the flash to make the overall effect more salient.