Door puzzle now has option to skip second part
The door puzzle now includes a bool, set in the inspector, that determines whether to skip over the second part of the puzzle (the part with the sequence that the player must repeat back to the gold instruments). It is set to false in the level, but may be enabled later, if this option is preferred.