Disable "get instrument" UI hint, update exhaust
The UI hint asking the player to find an instrument is no longer enabled. Instead, the ability icons are darkened until the player gets an instrument. Additionally, the percentage of the magic meter needed to play an instrument after completely exhausting the meter now has a control variable that can be accessed in the editor (Break Exhaust At Percent). The ability to play is now restored at this percentage recovered instead of only at 100%. Currently, it is set at 25%.