Replace enemies with robots temporarily
Enemies now use placeholder character models and animations (the prefabs still contain the original models, but they are deactivated).
Note that scene overrides, like the drunk enemy at the beginning, will retain the old models and need to be updated in the scene itself. The placeholders lack some animations. For example, an enemy charging a fireball will not show an animation in its current state and the doubt and hurt animations use variations of crouching to differentiate the states from the idle state. Additionally, reimported the Kenney character assets into a smaller selection, as the whole folder's worth of material was not in use except for a few assets used in this update. This should make for easier navigation and less clutter. Character assets that are needed later can always be reimported back into the folder again.