Bugfix puzzle and improve puzzle-player feedback
Fixed an issue with multiple puzzle hint loops running at once as well as an issue where the hinting loop would start at the wrong step after a wrong answer (which would confuse players as to the order unless they waited for the hint to restart several seconds later). Particles now properly display different colours depending on whether the player's input is correct or incorrect while solving the puzzle. Now, green particles should be displayed when hitting a statue in the right order, or all of them at once if the puzzle is completely solved. Additionally, red particles are emitted from all statues if the sequence was incorrect. White particles are still used for hinting the sequence between player attempts to solve the puzzle.