Add fixed version of river outside of main scene
Added an updated version of the first section of the island, packaged as a prefab to allow for easier changes to minor things related to the major level meshes when added to the main scene at a later date. This version of the island splits the river into multiple meshes to solve an issue with mesh colliders in the old version. Everything else in the model, including the UVs, are unchanged. Note that the new version is a different asset, so as not to override the existing asset that is currently in use.