Replace placeholder goal w/ violin, fix Bob/Rotate
The new violin model replaces the golden ukulele placeholder for the win condition pickup. Fixed a bug where the Bob/Rotate script was calculating the height of each bob with the variable for rotations per second instead of modifying the rotation speed like it was supposed to. Also made the violin rotate a bit slower so that players can get a proper look at its details.