Add moving, damageable forklift that drops violin
Added a forklift prefab that uses a nav mesh agent for pathfinding between nodes. The forklift can be damaged by the player and drops the violin upon being destroyed (default health is set at 1, but it can be set higher). Materials used for the dissolve effect on the forklift are instances of the dissolve material, which seems to avoid overwriting the properties of the reference material they are copied from as with the method of access that the other enemies use. It might be beneficial to implement this access method with the other enemies. Additionally, cleaned up code related to damage and health so that alternate hierarchy setups, as seen with the moving forklift prefab, work (ex. the damage/knockback script can be attached to a game object at a different level in the hierarchy than that of the hitbox). Removed the ability for players to exit out to the title screen by pressing Escape. This should prevent players from accidently losing progress if they start pressing keys randomly to see if they do anything.