Add trigger prefab to activate forklift
Added a trigger prefab to activate the forklift (or any other object(s) that are referenced by the script) when the player enters its area. Set up code for the secondary dissolving effect on forklift destruction.

Fixed an issue with the forklift's dissolvable models being offset from their base model.

Known issue(s):
The secondary dissolving effect upon forklift destruction seems to be running as expected, but it seems to be invisible when destruction effects play. No errors are thrown in regards to the effect's activation.