Add barrel pallet and violin onto forklift prefab
The forklift prefab now has the violin model visibly sitting on a pallet of barrels. Particle effects on the pallet are meant attract players towards breaking the forklift to get the violin. Additionally, the violin now uses the same particle effects as the message bottles instead of the player respawn particles.