Fix beach navmesh and add forklift in start area
The beach's stairs now are now properly included in navmesh bakes. Overrides on other stairs make it so enemies still will not exit certain areas. Enemies no longer consider the river to be part of the walkable navmesh. Additional camera posing, collisions, and forklift pathing added at beach, and the beach forklift has more health so that players do not rush through enemies to destroy it immediately (forklift has a health bar to show progressing damage). Added a forklift to the starting area, which is invulnerable and zooms out of sight with the violin as soon as the player sees it. Colliders on bridges have had their mesh renderers disabled so they do not shade the visible parts of the bridges. Disabled a checkpoint near the beach in the path of enemies meant to defend the forklift (the last checkpoint before it should suffice). Added more consistent colliders to the cliff over the beach to prevent sequence breaking, adding more wooden barricades to show this visually. Set player's rigid body component to interpolate instead of extrapolate. This will hopefully make collisions not as choppy. Disabled a ridge of rocks that had no collision and did not make sense in terms of the island no longer being attached to other terrain. To avoid players missing the ukulele pickup, it has been temporarily moved to the end of the first bridge and made unavoidable so players will always pick it up. Filled in a hole in the first bridge's collision. Moved the trigger that starts the oil drinking robot's movement towards the pitfall to its proper place (at the end of the bridge) and disabled an old child object that was preventing the script controlling the whole sequence from targeting the correct animator component.

Fixed a bug that made forklifts displace their models if using the oscillating movement script to make their model bob up and down. Oscillating movement on forklift models has been reimplemented as a result. Fixed a bug causing enemies to start emitting sparks instead of stopping whenever they started moving towards a new patrol node.