Fix inconsistent crash from damaging some enemies
Reverted a change to code handling player attack damage that broke some enemies depending on their position in the hierarchy. Forklifts now have a replacement hitbox at the top of the prefab's hierarchy, as the reverted change to the code was intended to allow the forklift prefabs to use colliders of children for their hitboxes.