Fix forklift health bug and fireball collisions
Fixed a bug in the forklift behaviour script that caused the script to crash if it reached the final node in its queue, rendering it unable to be damaged and softlocking the player's progress. Changed some settings in the fireball prefab to disable collisions with triggers, which was causing them to disappear before hitting anything "physical". Enemies can now take damage from fireballs and fireballs are set to no longer pass through players. Pitfall monsters now all have consistent vertical displacement from their code. Some remaining floating lanterns were grounded and an unintended tutorial bottle at the beach was disabled. The forklift at the start hovers over the stage instead of inside it. Forklift path at beach was modified and some trees removed so they are not growing on top of the path on the beach. Camera angles on the beach were modified to better display important elements. The violin's pickup range has been decreased so that players do not immediately pick it up and miss the chance to look at it once the forklift drops it.